build: add lombok dependency, change package structure
